refactor(compatibility): update for Stardew Valley 1.6 API changes
- Replace `Game1.log` with `Console.WriteLine` for logging due to access level changes
- Replace `crop.RegrowAfterHarvest` with `GetCropRegrowDays` method for fetching regrow days
- Update `crop.netSeedIndex` to `crop.indexOfHarvest` and handle type conversion with `int.Parse`
- Fix `Game1.player.maxStamina` usage to handle `NetInt` type
- Update `farm.buildStructure` to include new parameter for building placement
- Remove optional `LogLevel` parameter from `monitor.Debug` calls
- Refactor `CurrentPlayerLimit` to access the property directly
- Add exception handling for crop regrow days and season data retrieval

+++ MultiplayerAssistant/ModEntry.cs
@@ -55,8 +55,7 @@ namespace MultiplayerAssistant
             if (Context.IsWorldReady)
             {
                 Game1.player.health = Game1.player.maxHealth;
-                // 1.6 中 maxStamina 变为 NetInt，需要取 Value 再转换为 float
-                Game1.player.stamina = (float)Game1.player.maxStamina.Value;
+                Game1.player.stamina = Game1.player.maxStamina.Value; // 使用 .Value 获取 NetInt 的实际值
             }
         }
